IR_Drop/requirements.txt
datasets==3.1.0
diffusers==0.32.2
matplotlib==3.9.2
mmcv==1.3.11
numpy==2.0.0
opencv-python==4.10.0.84
pandas==2.2.3
pillow==10.4.0
psutil==6.0.0
pytorch-msssim==1.0.0
safetensors==0.4.5
SciencePlots==2.1.1
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
timm==1.0.14
torch==2.4.1
tqdm==4.67.0
transformers==4.46.2